requirements.txt
dash
pandas
plotly
requests
python-dotenv
dash-mantine-components
dash-bootstrap-components
dash-ag-grid